+++ src/app/cli/src/app.rs
@@ -884,11 +884,22 @@ pub fn register_config_in_catalog(
         embeddings_chunker,
         embeddings_encoder,
         vector_repo,
+        overfetch_factor,
+        overfetch_amount,
     } = config.search.clone().unwrap();
 
     catalog_builder.add::<kamu_search_services::SearchServiceLocalImplLazyInit>();
+    catalog_builder.add_value(kamu_search_services::SearchServiceLocalConfig {
+        overfetch_factor: overfetch_factor.unwrap(),
+        overfetch_amount: overfetch_amount.unwrap(),
+    });
+
+    let indexer = indexer.unwrap_or_default();
     catalog_builder.add_value(kamu_search_services::SearchServiceLocalIndexerConfig {
-        clear_on_start: indexer.unwrap_or_default().clear_on_start,
+        clear_on_start: indexer.clear_on_start,
+        skip_datasets_with_no_description: indexer.skip_datasets_with_no_description,
+        skip_datasets_with_no_data: indexer.skip_datasets_with_no_data,
+        payload_include_content: indexer.payload_include_content,
     });
 
     match embeddings_chunker.unwrap_or_default() {