+++ src/app/cli/src/commands/inspect_schema_command.rs
@@ -119,10 +119,10 @@ impl InspectSchemaCommand {
         match e {
             QueryError::DatasetNotFound(e) => CLIError::usage_error_from(e),
             QueryError::DatasetSchemaNotAvailable(_) => unreachable!(),
-            e @ (QueryError::DatasetBlockNotFound(_)
-            | QueryError::DataFusionError(_)
-            | QueryError::Access(_)) => CLIError::failure(e),
-            e @ QueryError::Internal(_) => CLIError::critical(e),
+            QueryError::DatasetBlockNotFound(_)
+            | QueryError::BadQuery(_)
+            | QueryError::Access(_) => CLIError::failure(e),
+            QueryError::Internal(_) => CLIError::critical(e),
         }
     }